requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
docker==6.1.3
requests==2.31.0
PyYAML==6.0.1
psutil==5.9.5